Shuffle round

+++ docker_entry.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
+# quick self-identity check
+grep -cm1 docker /proc/1/cgroup 2> /dev/null || echo "* Cannot execute outside Docker" && exit 1
 echo "* Running coverage checks..."
 rm -rf .cov_temp
 istanbul cover --config .istanbul-config.yml test/test.js
